0.7 work continuing. Added basic logic to the command_infrastructure.go, added the /age command
