+++ command/command.go
@@ -259,3 +259,60 @@ func SendTextEmbed(s *discordgo.Session, m *discordgo.MessageCreate, status stri
 		return
 	}
 }
+func SendTextEmbedCommand(s *discordgo.Session, m string, status string, messageContent string, mode discordgo.EmbedType) {
+
+	//Fixed author message.
+	author := discordgo.MessageEmbedAuthor{
+		URL:          "",
+		Name:         "SlovakiaBot",
+		IconURL:      "https://cdn.discordapp.com/avatars/837982234597916672/51236a8235b1778f5d90bce35fbcf4d6.webp?size=256",
+		ProxyIconURL: "",
+	}
+
+	var color int
+
+	//set the embed color according to the type of Status passed (OK, ERR, WARN, SYNTAX)
+	switch status {
+	case ":bangbang: ERROR":
+		color = 15158332
+	case ":warning: WARNING":
+		color = 15105570
+	case ":question: SYNTAX":
+		color = 3447003
+	case ":no_entry: AUTHENTICATION":
+		color = 15105570
+	case ":wrench: AUTOCORRECTING":
+		color = 16776960
+
+	default:
+		color = 3066993
+	}
+
+	var embedArray []discordgo.MessageEmbed
+	//MessageEmbed info
+	//Thinking of adding timestamp time.Now().Format(time.RFC3339)
+	embed := discordgo.MessageEmbed{
+		URL:         "",
+		Type:        mode,
+		Title:       status,
+		Description: messageContent,
+		Timestamp:   "",
+		Color:       color,
+		Footer:      nil,
+		Image:       nil,
+		Thumbnail:   nil,
+		Video:       nil,
+		Provider:    nil,
+		Author:      &author,
+		Fields:      nil,
+	}
+
+	embedArray = append(embedArray, embed)
+
+	//Send a message as an embed.
+	_, err := s.ChannelMessageSendEmbed(m, &embed)
+	if err != nil {
+		s.ChannelMessageSend(m, "**[TESTERR]** Error: "+err.Error())
+		return
+	}
+}
